+++ CHANGES.md
@@ -1,3 +1,7 @@
+## 1.1.0 (2013/?/?)
+ * Feature #8: Allow client session id to be specified. If not
+   specified, auto-generate one.
+
 ## 1.0.0 (2013/7/17)
  * Feature #7: Permissions can be allowed in bulk for all topics, or
    per topic category.